refactor(search): move cache and its creation in separate file
