+++ scraper.ts
@@ -8,7 +8,7 @@ import { resolve as resolvePath, join as joinPath } from 'path';
 import { spawn } from 'child_process';
 import Trie from 'compact-prefix-tree/cjs';
 import { SUPPORTED_TYPES, DATA_DIR } from './constants.js';
-import { Data } from './search.js';
+import { Data } from './cache.js';
 
 const { readdir, readFile, writeFile } = fs;
